requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
eventlet==0.33.3
requests==2.31.0
flask-cors==4.0.0  # Only needed if frontend served from different port
